Commentary review changes
