+++ src/components/read/BiblePane.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import { Swipeable } from "react-swipeable";
+// import { Swipeable } from "react-swipeable";
 import Grid from "@material-ui/core/Grid";
 import Fullscreen from "react-full-screen";
 import MenuBar from "./MenuBar";
@@ -26,35 +26,38 @@ const BiblePane = ({ setValue, paneData, ref1, scroll, paneNo }) => {
   const classes = useStyles();
   const [fullscreen, setFullscreen] = React.useState(false);
   //edit this function to show the menu at a later stage
-  const showMenu = () => console.log("showMenu");
+  //const showMenu = () => console.log("showMenu");
   //edit this function to hide the menu at a later stage
-  const hideMenu = () => console.log("hideMenu");
+  // const hideMenu = () => console.log("hideMenu");
   return (
-    <Swipeable onSwipedUp={hideMenu} onSwipedDown={showMenu}>
-      <MenuBar
-        {...paneData}
-        setFullscreen={setFullscreen}
-        setValue={setValue}
-      />
-      <Grid container className={classes.bible}>
-        <Grid item xs={12}>
-          <Fullscreen
-            enabled={fullscreen}
-            onChange={fullscreen => setFullscreen(fullscreen)}
-            className={classes.fullscreen}
-          >
-            <Bible
-              {...paneData}
-              setValue={setValue}
-              ref1={ref1}
-              scroll={scroll}
-              paneNo={paneNo}
-            />
-          </Fullscreen>
+    <>
+      {/* <Swipeable onSwipedUp={hideMenu} onSwipedDown={showMenu}> */}
+      <div>
+        <MenuBar
+          {...paneData}
+          setFullscreen={setFullscreen}
+          setValue={setValue}
+        />
+        <Grid container className={classes.bible}>
+          <Grid item xs={12}>
+            <Fullscreen
+              enabled={fullscreen}
+              onChange={fullscreen => setFullscreen(fullscreen)}
+              className={classes.fullscreen}
+            >
+              <Bible
+                {...paneData}
+                setValue={setValue}
+                ref1={ref1}
+                scroll={scroll}
+                paneNo={paneNo}
+              />
+            </Fullscreen>
+          </Grid>
         </Grid>
-      </Grid>
-    </Swipeable>
+      </div>
+      {/* </Swipeable> */}
+    </>
   );
 };
-
 export default BiblePane;
