+++ src/store/api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
 export default axios.create({
-  baseURL: "https://api.vachanonline.net/v1",
+  baseURL: "http://127.0.0.1:8000/v1",
   timeout: 15000
 });
